+++ CHANGELOG.md
@@ -2,6 +2,7 @@ XXX
 ===================
 - Revereted back - `Apify.getApifyProxyUrl()` accepts again `session` and `groups` options instead of
   `apifyProxySession` and `apifyProxyGroups`
+- Parameter `memory` added to `Apify.call()`.
 
 0.5.37 / 2018-06-07
 ===================